changed as.integer to as.numeric

+++ R/LearnerSurvCoxTimeSem.R
@@ -152,7 +152,7 @@ LearnerSurvCoxtime2 = R6::R6Class("LearnerSurvCoxtime2",
                                      # Set-up network architecture
                                      
                                      # num_nodes needs to be reconstructed
-                                     num_nodes <- rep(as.integer(self$param_set$get_values(tags = "net")$nodes_per_layer),
+                                     num_nodes <- rep(as.numeric(self$param_set$get_values(tags = "net")$nodes_per_layer),
                                                       self$param_set$get_values(tags = "net")$num_layers)
                                      
                                      # num_nodes_raw = c(self$param_set$get_values(tags = "net")$num_nodes1,
